requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==5.2.13
django-cors-headers==4.9.0
django-taggit==6.1.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
pillow==12.2.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
django-extensions==4.1
django-anymail==14.0
pillow==12.2.0
certifi==2026.2.25
charset-normalizer==3.4.7
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
idna==3.11
requests==2.33.1
six==1.17.0
urllib3==2.6.3